requirements.txt
# Базовые зависимости
python-dotenv==1.0.0
pytest==7.4.0
loguru==0.7.0

# Работа с Neo4j
neo4j>=5.12.0

# OpenAI и API интеграции
openai>=1.7.2
httpx>=0.24.1

# Telegram бот
python-telegram-bot>=20.6

# Веб-интерфейс
fastapi>=0.103.1
uvicorn>=0.23.2
jinja2>=3.1.2

# Утилиты
pydantic>=2.4.0
